+++ e2e/fixtures/fixtures.js
@@ -42,7 +42,12 @@ describe('foo', () => {
     expect(foo())${testShouldPass ? '' : '.not'}.to.equal('got is-type and got is-string and got foo');
   });
 });`;
+export const appPrintIsType = "const isType = require('./components/utils/is-type'); console.log(isType());";
+export const appPrintIsTypeCapsule = "const isType = require('.'); console.log(isType());";
+export const appPrintIsString = "const isString = require('./components/utils/is-string'); console.log(isString());";
+export const appPrintIsStringCapsule = "const isString = require('.'); console.log(isString());";
 export const appPrintBarFoo = "const barFoo = require('./components/bar/foo'); console.log(barFoo());";
+export const appPrintBarFooCapsule = "const barFoo = require('.'); console.log(barFoo());";
 export const appPrintBarFooES6 = "const barFoo = require('./components/bar/foo'); console.log(barFoo.default());";
 export const appPrintBarFooAuthor = "const barFoo = require('./bar/foo'); console.log(barFoo());";
 export const objectRestSpread = `const g = 5;