+++ e2e/flows/copy-workspace-with-env.e2e.2.js
@@ -6,13 +6,13 @@ import { DIAGNOSIS_NAME } from '../../src/doctor/core-diagnoses/broken-symlink-f
 
 chai.use(require('chai-fs'));
 
-describe('custom module resolutions', function () {
+describe('copy workspace with env', function () {
   this.timeout(0);
   const helper = new Helper();
   after(() => {
     helper.destroyEnv();
   });
-  describe('using custom module directory', () => {
+  describe('coping a workspace with a compiler to another directory', () => {
     let copiedPath;
     before(() => {
       helper.setNewLocalAndRemoteScopes();