+++ src/cli/commands/private-cmds/ci-update-cmd.js
@@ -15,6 +15,7 @@ export default class CiUpdate extends Command {
     // ['v', 'verbose [boolean]', 'showing npm verbose output for inspection'],
     ['d', 'directory [file]', 'directory to run ci-update'],
     ['k', 'keep', 'keep test environment after run (default false)'],
+    ['c', 'no-cache', 'ignore component cache when creating dist file'],
     ['o', 'output [file]', 'save ci results to file system']
   ];
   /* changed command to not private so bit will not print {"payload":"","headers":{"version":"0.12.10-dev.1"}} when it fails */
@@ -26,16 +27,18 @@ export default class CiUpdate extends Command {
       // verbose = true,
       directory,
       output,
-      keep = false
+      keep = false,
+      noCache = false
     }: {
       // verbose: ?boolean,
       directory: ?string,
       output: ?string,
-      keep: boolean
+      keep: boolean,
+      noCache: boolean
     }
   ): Promise<any> {
     const verbose = true; // During ci-update we always want to see verbose outputs
-    return ciUpdateAction(id, scopePath || process.cwd(), verbose, directory, keep).then(
+    return ciUpdateAction(id, scopePath || process.cwd(), verbose, directory, keep, noCache).then(
       ({ specsResults, dists, mainFile }) => ({ specsResults, dists, mainFile, output, directory })
     );
   }