+++ src/consumer/component-ops/many-components-writer.js
@@ -9,7 +9,7 @@ import { installNpmPackagesForComponents } from '../../npm-client/install-packag
 import * as packageJsonUtils from '../component/package-json-utils';
 import type { ComponentWithDependencies } from '../../scope';
 import type Component from '../component/consumer-component';
-import { COMPONENT_ORIGINS } from '../../constants';
+import { COMPONENT_ORIGINS, DEFAULT_DIR_DEPENDENCIES } from '../../constants';
 import logger from '../../logger/logger';
 import type Consumer from '../consumer';
 import { isDir, isDirEmptySync } from '../../utils';
@@ -174,7 +174,10 @@ export default class ManyComponentsWriter {
     );
   }
   _getWriteParamsOfOneComponent(componentWithDeps: ComponentWithDependencies): ComponentWriterProps {
-    const componentRootDir: PathOsBasedRelative = this._getComponentRootDir(componentWithDeps.component.id);
+    // for isolated components, the component files should be on the root. see #1758
+    const componentRootDir: PathOsBasedRelative = this.isolated
+      ? '.'
+      : this._getComponentRootDir(componentWithDeps.component.id);
     const getParams = () => {
       if (!this.consumer) {
         return {
@@ -323,7 +326,7 @@ export default class ManyComponentsWriter {
   }
   _getDependencyRootDir(bitId: BitId): PathOsBasedRelative {
     if (this.isolated) {
-      return composeDependencyPathForIsolated(bitId);
+      return composeDependencyPathForIsolated(bitId, DEFAULT_DIR_DEPENDENCIES);
     }
     // $FlowFixMe consumer is set here
     return this.consumer.composeRelativeDependencyPath(bitId);